move functions into cpp files
